Updates to project template for 0.12.0
The plan will be to merge these changes to the release-0.12.0 branch
once ready. Before releasing RapidSMS 0.12.0, tag the tip of this
release-0.12.0 branch as '0.12.0' so that users can choose this
version specifically for use with RapidSMS 0.12.0.  Then merge all
the changes to master and continue development there.

Branch: update-for-0.12.0